fix: dnd 計算邏輯、遇到小數時的邏輯尚未處理好

+++ taskify-backend/src/main/java/com/twoyu/taskifybackend/model/entity/StatusColumn.java
@@ -46,7 +46,7 @@ public class StatusColumn implements Serializable {
      * 狀態欄排序
      */
     @Column(name = "data_index", nullable = false)
-    private Integer dataIndex;
+    private Double dataIndex;
 
     @Column(name = "board_id", nullable = false)
     private UUID boardId;
