+++ taskify-backend/src/main/java/com/twoyu/taskifybackend/model/vo/response/UpdateColumnResponse.java
@@ -2,5 +2,5 @@ package com.twoyu.taskifybackend.model.vo.response;
 
 import java.util.UUID;
 
-public record UpdateColumnResponse(UUID id, UUID boardId, String title, int dataIndex) {
+public record UpdateColumnResponse(UUID id, UUID boardId, String title, Double dataIndex) {
 }
